// backend/src/models/types/property-type.ts
export default interface PropertyType {

	name: string
	username: string
	logo: string
	city: string
	state: string
	country: string
	property_type: string
	longitude: string
	latitude: string
	email: string
	taxid: string
	taxverified: string
	website: string
	company_id: string
	number_of_rooms: string
	phone: number
	timezone: string
	postal_code: string
	selling_date: string
	api_key: string

}